Tests: Move extension method
